+++ gitRemoteAws/botoman.py
@@ -7,6 +7,8 @@ import yaml
 import os
 from botocore.exceptions import ProfileNotFound
 import sys
+import logging
+logger = logging.getLogger("git-remote-aws")
 
 class SessionMan:
     """
@@ -29,12 +31,12 @@ class SessionMan:
         try:
             return boto3.session.Session(profile_name=profile_name)
         except ProfileNotFound as error:
-            sys.stderr.write("fatal: %s"%str(error))
+            logger.error("fatal: %s"%str(error))
 
             # if not at clone stage
             if self.dm is not None:
-                sys.stderr.write("Perhaps edit the 'profile' field in %s"%self.dm.fn['config'])
-                sys.stderr.write("to match with the profiles in %s"%self.dm.fn['aws_credentials'])
+                logger.error("Perhaps edit the 'profile' field in %s"%self.dm.fn['config'])
+                logger.error("to match with the profiles in %s"%self.dm.fn['aws_credentials'])
 
             sys.exit(1)